wip: refactoring camera and movement scripts
