--- Assets/Resources/Models/Interactive Models/Sci-fi Doorway 01/Materials.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4d600ddbfde1db144aef05b129d728c3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
