--- Assets/Resources/Models/Interactive Models/Sci-fi Doorway 01/Prefabs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3be59088f5d8d5f4f9f8617f682f00ef
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
